Add child combinator
This is the implementation of the child combinator in `Floki.find/2`.
It is needed when the user only wants to select the children
elements of a parent, and not every descendent node.

Examples of query:

- Floki.find(html, "div > a")
- Floki.find(html, ".container > a > img")

Specs at W3C: http://www.w3.org/TR/css3-selectors/#child-combinators

This commit closes #23