Handle boolean attributes with raw_html/1
A boolean attribute is one who's lack of value indicates truthiness.
Since a boolean attribute does not have a value, it doesn't fit into the
assumption that each attribute entry in the AST is a 2 element tuple.
Therefore the rendering function is updated to handle non-tuple
attributes as if boolean attributes.

Note: This AST form, e.g. {"div", ["hidden"], []}, is not a valid AST
produced by mochiweb_html (which is the parser used internally by
Floki). I thought this would be OK since the fact that such a parser is
being used is an implementation detail. Producing HTML _from_ the AST is
a different responsibility, which is in fact owned by the Floki lib.