Add HTML tree: an abstraction to represent a tree with a map
It keeps references for each node and the relationship between them.
This is needed in order to enable update the tree. It also enables
analyse the tree after building it. For example, it's possible to lookup
the parent node of a given node, and check the position of that given
node related to parent.