Refactor in Floki.Finder to use HTML Tree for traverse and search
This is needed in order to be the base of the next features
such as: update tree, find nodes using the "pseudo class" :nth-child,
filter out elements with selectors, and more.

With this refactor it is now easier to analyse the position and to get
nodes that are related to a given node.

It closes issue #71 - https://github.com/philss/floki/issues/71.