Refactor Floki.Finder to use HTML Tree for traverse and search (#73)
* Fix async for ExUnit tests

* Add HTML tree: an abstraction to represent a tree with a map

It keeps references for each node and the relationship between them.
This is needed in order to enable update the tree. It also enables
analyse the tree after building it. For example, it's possible to lookup
the parent node of a given node, and check the position of that given
node related to parent.

* Comment why there is no docs for the function

* Refactor in Floki.Finder to use HTML Tree for traverse and search

This is needed in order to be the base of the next features
such as: update tree, find nodes using the "pseudo class" :nth-child,
filter out elements with selectors, and more.

With this refactor it is now easier to analyse the position and to get
nodes that are related to a given node.

It closes issue #71 - https://github.com/philss/floki/issues/71.