Remove documentation from modules and functions that are private
This is needed to make the documentation cleaner and avoid usage of
functions that are "internal".

It's better to have code comments instead of public documentation for
internal code, since only programmers of this library should worry about
its implementation details.

This way programmers and users can focus on what is public or not.