Fix ":not()" pseudo class selector to accept pseudo-class selectors as args (#104)
* Enable selector parser to understand pseudo class inside :not

To make the Finder work properly, we need to make the Selector
understand about pseudo classes, in order to work recursively.

* Move pseudo-class match to inside selector

This is because we need recursion when matching selectors inside the
"not" pseudo-class selector. It make possible to use an expression like
this: `a:not(:nth-child(2))`. For this to work, the tree had to be
passed to the selector as well.

* Add test case for :not() with other pseudo-class

* Remove unnecessary case because it will always find the node in that point