Replace parse/1 with parse_document/1 parse_fragment/1
This commit deprecates the Floki.parse/1 function and functions that
uses it under the hood, and suggest the usage of Floki.parse_document/1
or Floki.parse_fragment/1.

This is because we get different trees if we choose for parsing using
the fragment algorithm vs the document algorithm. With the introduction
of `fast_html`, this became more clear that affects the output.

The idea for those functions is to return an Either instead of the
document directly. This is because it is possible that the parser
returns errors, like the module not being loaded or a problem with the
input itself. Also, the document returned will be always a list.

For this refactor I had to improve a little bit the tests. The most
important thing we had to do is to parse the documents before using with
functions like `Floki.find/2` or `Floki.attribute/3`. It is more easier
to create setups now, with the `document!/1` and `html_body/1`
functions.