Change parser dependencies to be more explict
Since html5ever is not working with OTP 22, I need to test it against
the GitHub version. To make more easier to configure that, I changed the
dependencies declaration to be more explicit.

Also, only "fast_html" and "html5ever" packages should be added, since
"mochiweb" should not be added as dependency anymore.