Add Floki.Traversal.traverse_and_update/3
This function enables updating the tree with an accumulator. It is
useful in cases where you need to keep track of changes.

With this function, we can have a functional approach to
traversing and manipulating HTML trees with a state.

This commit also improves `Floki.Traversal`, and improves the docs for
`Floki.traverse_and_update/2` (the one without accumulator).

Closes https://github.com/philss/floki/issues/246