Add nth-last-of-type pseudo-class selector
This selector is similar to "nth-of-type", except it counts items
backwards from the end, not forwards from the beginning.

This also refactor the pseudo class module to remove some functions in
order to be more explicit in the main function bodies.

It uses this pseudo class internally to fetch the "last-of-type".

Reference: https://developer.mozilla.org/en-US/docs/Web/CSS/:nth-last-of-type