Use hexpm Docker images for our CI workflows
It replaces the "official" imagens with images provided by the Hex team,
which are updated automatically for each branch and tag created on the
Elixir project. They are probably more stable and are released faster
than the official channel.

For more details see https://elixirforum.com/t/yet-another-elixir-and-erlang-docker-image/28740
And also check https://github.com/hexpm/bob