:pager: Configure `parser` @tag to skip tests
Due to some ExUnit tag limitations we can not use `@tag parser [Parser1, Parser2]`. (ExUnit does not allow tags to be list if they are excluded/included)
That's why in the `except_parser` tag was created to run a test for all parsers except a specific one. This works fine when we have only 3 parsers...

if in the future Floki has more than 3 parsers and a test needs to include only 2, I think the best approach would be to create a test macro that will be used like:

```elixir
test_for_parsers [Parser1, Parser2] do
  ...
end
```