Parse `title` tag as plaintext with mochiweb and add `parser` ExUnit tags (#343)
* :pager: Configure `parser` @tag to skip tests

Due to some ExUnit tag limitations we can not use `@tag parser [Parser1, Parser2]`. (ExUnit does not allow tags to be list if they are excluded/included)

That's why in the `except_parser` tag was created to run a test for all parsers except a specific one. This works fine when we have only 3 parsers.

* :turtle: Mochiweb - parse `title` tag as plaintext