Add a HTML tokenizer written in Elixir (#204)
This is part of a bigger effort to write a compliant HTML parser in
Elixir.

The implementation follows WHATWG specification which is the living
standard of HTML, but parts of the tokenizer are still missing like the
handling of parse errors and some states. Those missing parts are not
essential for most of the documents.

You can see details about the HTML specification here:
https://html.spec.whatwg.org/multipage/

This commit contains a lot of files. The most important one is the
`lib/floki/html/tokenizer.ex`. We added a lot of test files that were
generated according to html5lib-tests - a project that aims to provide
test cases based on WHATWG specs.
See: https://github.com/html5lib/html5lib-tests

This tokenizer was written based on the specs as seen around September
2019. Most of the parser development progress is being tracked at
https://github.com/philss/floki/projects/2

For now it will remain "private" and no other module is using it.

This is related to https://github.com/philss/floki/issues/37 :)