Fix pseudo-class ":not" selector parsing halting point
This is a fix for when a "pseudo-class" ":not" that contains an
attribute selector is followed by another selector.
This is an example: "a:not([class]), div".

The solution was to avoid using the general parsing state for the
attribute, and instead use the "consume attribute" state, which
is more predictable.

Closes https://github.com/philss/floki/issues/441